Add a ConfigureSharedCookie helper to frameworks
