+++ samples/MvcApp/packages.config
@@ -16,10 +16,10 @@
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net472" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net472" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.Cryptography.Internal" version="6.0.7" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.DataProtection" version="6.0.7" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.DataProtection.Abstractions" version="6.0.7" targetFramework="net472" />
-  <package id="Microsoft.AspNetCore.DataProtection.Extensions" version="6.0.7" targetFramework="net472" />
+  <package id="Microsoft.AspNetCore.Cryptography.Internal" version="6.0.8" targetFramework="net472" />
+  <package id="Microsoft.AspNetCore.DataProtection" version="6.0.8" targetFramework="net472" />
+  <package id="Microsoft.AspNetCore.DataProtection.Abstractions" version="6.0.8" targetFramework="net472" />
+  <package id="Microsoft.AspNetCore.DataProtection.Extensions" version="6.0.8" targetFramework="net472" />
   <package id="Microsoft.AspNetCore.Hosting.Abstractions" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.AspNetCore.Hosting.Server.Abstractions" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.AspNetCore.Http.Abstractions" version="2.1.1" targetFramework="net472" />
@@ -35,11 +35,11 @@
   <package id="Microsoft.Extensions.Options" version="6.0.0" targetFramework="net472" />
   <package id="Microsoft.Extensions.Primitives" version="6.0.0" targetFramework="net472" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.11" targetFramework="net472" />
-  <package id="Microsoft.Owin" version="4.2.1" targetFramework="net472" />
-  <package id="Microsoft.Owin.Host.SystemWeb" version="4.2.1" targetFramework="net472" />
-  <package id="Microsoft.Owin.Security" version="4.2.1" targetFramework="net472" />
-  <package id="Microsoft.Owin.Security.Cookies" version="4.2.1" targetFramework="net472" />
-  <package id="Microsoft.Owin.Security.Interop" version="2.1.2" targetFramework="net472" />
+  <package id="Microsoft.Owin" version="4.2.2" targetFramework="net472" />
+  <package id="Microsoft.Owin.Host.SystemWeb" version="4.2.2" targetFramework="net472" />
+  <package id="Microsoft.Owin.Security" version="4.2.2" targetFramework="net472" />
+  <package id="Microsoft.Owin.Security.Cookies" version="4.2.2" targetFramework="net472" />
+  <package id="Microsoft.Owin.Security.Interop" version="2.2.0" targetFramework="net472" />
   <package id="Microsoft.Owin.Security.OAuth" version="4.2.1" targetFramework="net472" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net472" />
   <package id="Microsoft.Win32.Registry" version="5.0.0" targetFramework="net472" />
@@ -53,7 +53,7 @@
   <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" />
   <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.AccessControl" version="6.0.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Xml" version="6.0.0" targetFramework="net472" />
+  <package id="System.Security.Cryptography.Xml" version="6.0.1" targetFramework="net472" />
   <package id="System.Security.Permissions" version="4.5.0" targetFramework="net472" />
   <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net472" />
   <package id="System.Text.Encodings.Web" version="6.0.0" targetFramework="net472" />
